The next tasks are ready: 9. Palindrome Number 12. Integer to Roman 13. Roman to Integer 14. Longest Common Prefix 28. Find the Index of the First Occurrence in a String 45. Jump Game II 58. Length of Last Word 125. Valid Palindrome 200. Number of Islands 242. Valid Anagram 392. Is Subsequence

+++ src/55_jump_game.cpp
@@ -12,7 +12,7 @@ namespace _55 {
     * O(1)
     */
     bool Solution::canJump(std::vector<int>& nums) {
-        int n = nums.size();
+        size_t n = nums.size();
         if (n <= 1) {
             return true;
         }
